Organize files into appropriate directories
Fix issues with test not running
Complete 001-Applications issue ready for PR
