+++ src/testUserContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const testUserContext = createContext();
+
+export default testUserContext;
